requirements.txt
requests
boto3
openai
moviepy==2.0.0.dev2
opencv-python
numpy
Pillow
python-dotenv